Rename bot as rivescript
