+++ config/lib/rivescript.js
