+++ lib/helpers.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const logger = require('heroku-logger');
-const botConfig = require('../config/lib/bot');
+const botConfig = require('../config/lib/rivescript');
 
 /**
  * Returns whether given botReply should be handled via macro.
