+++ lib/middleware/receive-message/reply-brain.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const bot = require('../../bot.js');
+const rivescript = require('../../rivescript.js');
 
-module.exports = function getBotReply() {
+module.exports = function getRivescriptReply() {
   return (req, res, next) => {
-    bot.getReply(req.conversation, req.body.text)
+    rivescript.getReply(req.conversation, req.body.text)
       .then((reply) => {
         req.reply = reply;
