+++ lib/rivescript.js
@@ -3,7 +3,7 @@
 const RiveScript = require('rivescript');
 const logger = require('heroku-logger');
 const Campaigns = require('../app/models/Campaign');
-const config = require('../config/lib/bot');
+const config = require('../config/lib/rivescript');
 
 const defaultTopic = 'random';
